No CGO for rest consumers/clients
Refactored portions of the rest code to not require CGO for rest consumers.  Most of the refactored code was moved into pkg/rest/define to isolate it.

Also pruned an unused function and related unit test.
