+++ pkg/rest/config.go
@@ -1,10 +1,1 @@
 package rest
-
-type ServiceScheme int
-
-const (
-	TCP ServiceScheme = iota
-	Unix
-	None
-	HTTP
-)
