+++ pkg/rest/rest.go
@@ -1,60 +1,18 @@
 package rest
 
 import (
-	"errors"
-	"fmt"
 	"net/http"
-	"net/url"
-	"strings"
 
-	"github.com/crc-org/vfkit/pkg/cmdline"
 	"github.com/crc-org/vfkit/pkg/rest/define"
 	"github.com/gin-gonic/gin"
 	"github.com/sirupsen/logrus"
 )
 
-type Endpoint struct {
-	Host   string
-	Path   string
-	Scheme ServiceScheme
-}
-
-func NewEndpoint(input string) (*Endpoint, error) {
-	uri, err := parseRestfulURI(input)
-	if err != nil {
-		return nil, err
-	}
-	scheme, err := toRestScheme(uri.Scheme)
-	if err != nil {
-		return nil, err
-	}
-	return &Endpoint{
-		Host:   uri.Host,
-		Path:   uri.Path,
-		Scheme: scheme,
-	}, nil
-}
-
-func (ep *Endpoint) ToCmdLine() ([]string, error) {
-	args := []string{"--restful-uri"}
-	switch ep.Scheme {
-	case Unix:
-		args = append(args, fmt.Sprintf("unix://%s", ep.Path))
-	case TCP:
-		args = append(args, fmt.Sprintf("tcp://%s%s", ep.Host, ep.Path))
-	case None:
-		return []string{}, nil
-	default:
-		return []string{}, errors.New("invalid endpoint scheme")
-	}
-	return args, nil
-}
-
 // VFKitService is used for the restful service; it describes
 // the variables of the service like host/path but also has
 // the router object
 type VFKitService struct {
-	*Endpoint
+	*define.Endpoint
 	router *gin.Engine
 }
 
@@ -63,9 +21,9 @@ func (v *VFKitService) Start() {
 	go func() {
 		var err error
 		switch v.Scheme {
-		case TCP:
+		case define.TCP:
 			err = v.router.Run(v.Host)
-		case Unix:
+		case define.Unix:
 			err = v.router.RunUnix(v.Path)
 		}
 		logrus.Fatal(err)
@@ -75,7 +33,7 @@ func (v *VFKitService) Start() {
 // NewServer creates a new restful service
 func NewServer(vm *VzVirtualMachine, endpoint string) (*VFKitService, error) {
 	r := gin.Default()
-	ep, err := NewEndpoint(endpoint)
+	ep, err := define.NewEndpoint(endpoint)
 	if err != nil {
 		return nil, err
 	}
@@ -132,54 +90,4 @@ func (vm *VzVirtualMachine) setVMState(c *gin.Context) {
 		return
 	}
 	c.Status(http.StatusAccepted)
-}
-
-// parseRestfulURI validates the input URI and returns an URL object
-func parseRestfulURI(inputURI string) (*url.URL, error) {
-	restURI, err := url.ParseRequestURI(inputURI)
-	if err != nil {
-		return nil, err
-	}
-	scheme, err := toRestScheme(restURI.Scheme)
-	if err != nil {
-		return nil, err
-	}
-	if scheme == TCP && len(restURI.Host) < 1 {
-		return nil, errors.New("invalid TCP uri: missing host")
-	}
-	if scheme == TCP && len(restURI.Path) > 0 {
-		return nil, errors.New("invalid TCP uri: path is forbidden")
-	}
-	if scheme == TCP && restURI.Port() == "" {
-		return nil, errors.New("invalid TCP uri: missing port")
-	}
-	if scheme == Unix && len(restURI.Path) < 1 {
-		return nil, errors.New("invalid unix uri: missing path")
-	}
-	if scheme == Unix && len(restURI.Host) > 0 {
-		return nil, errors.New("invalid unix uri: host is forbidden")
-	}
-	return restURI, err
-}
-
-// toRestScheme converts a string to a ServiceScheme
-func toRestScheme(s string) (ServiceScheme, error) {
-	switch strings.ToUpper(s) {
-	case "NONE":
-		return None, nil
-	case "UNIX":
-		return Unix, nil
-	case "TCP", "HTTP":
-		return TCP, nil
-	}
-	return None, fmt.Errorf("invalid scheme %s", s)
-}
-
-func validateRestfulURI(inputURI string) error {
-	if inputURI != cmdline.DefaultRestfulURI {
-		if _, err := parseRestfulURI(inputURI); err != nil {
-			return err
-		}
-	}
-	return nil
 }
